pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "clean-room-agent"
version = "0.1.0"
description = "A self-improving coding agent built on context curation"
requires-python = ">=3.11"
dependencies = [
    "click>=8.0",
    "tree-sitter>=0.25,<0.26",
    "tree-sitter-python>=0.23",
    "tree-sitter-javascript>=0.23",
    "tree-sitter-typescript>=0.23",
    "httpx>=0.27",
    "pathspec>=0.12",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.8",
]

[project.scripts]
cra = "clean_room_agent.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/clean_room_agent"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
